Clean up some unused code
