+++ core/core/src/main/kotlin/edu/wpi/axon/core/dsl/CodeGraph.kt
@@ -125,5 +125,3 @@ class CodeGraph(
         return visited
     }
 }
-
-private infix fun <E> Iterable<E>.anyIn(other: Iterable<E>) = any { it in other }
