+++ core/core/src/main/kotlin/edu/wpi/axon/core/dsl/PolymorphicNamedDomainObjectContainerUtil.kt
@@ -1,0 +1,23 @@
+package edu.wpi.axon.core.dsl
+
+import edu.wpi.axon.core.dsl.container.PolymorphicNamedDomainObjectContainer
+import edu.wpi.axon.core.dsl.container.PolymorphicNamedDomainObjectContainerDelegateProvider
+import edu.wpi.axon.core.dsl.task.Task
+import edu.wpi.axon.core.dsl.variable.Variable
+import kotlin.reflect.KClass
+
+/**
+ * Creates a new variable and configures it.
+ */
+fun <T : Variable, U : T> PolymorphicNamedDomainObjectContainer<T>.creating(
+    type: KClass<U>,
+    configuration: (U.() -> Unit)? = null
+) = PolymorphicNamedDomainObjectContainerDelegateProvider.of(this, type, configuration)
+
+/**
+ * Creates a new task and configures it.
+ */
+fun <T : Task, U : T> PolymorphicNamedDomainObjectContainer<T>.running(
+    type: KClass<U>,
+    configuration: (U.() -> Unit)? = null
+) = PolymorphicNamedDomainObjectContainerDelegateProvider.of(this, type, configuration)
