+++ core/core/src/main/kotlin/edu/wpi/axon/core/dsl/task/InferenceTask.kt
@@ -9,7 +9,7 @@ import edu.wpi.axon.core.dsl.variable.Variable
 /**
  * A [Task] that runs inference.
  */
-class InferenceTask(override val name: String) : Task {
+class InferenceTask(name: String) : Task(name) {
 
     /**
      * The data input to the first layer of the model.
